+++ ionic-start/src/main.ts
@@ -1,13 +1,39 @@
 import { enableProdMode } from '@angular/core';
-import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
-
-import { AppModule } from './app/app.module';
 import { environment } from './environments/environment';
+import { bootstrapApplication } from '@angular/platform-browser';
+import { AppComponent } from './app/app.component';
+import {
+  provideVSCodeDesignSystem,
+  vsCodeButton,
+  vsCodeCheckbox,
+  vsCodeDivider,
+  vsCodeLink,
+  vsCodePanelView,
+  vsCodeProgressRing,
+  vsCodeRadio,
+  vsCodeRadioGroup,
+  vsCodeTag,
+  vsCodeTextField,
+} from '@vscode/webview-ui-toolkit';
 
 if (environment.production) {
   enableProdMode();
 }
 
-platformBrowserDynamic()
-  .bootstrapModule(AppModule)
-  .catch((err) => console.error(err));
+// In order to use the Webview UI Toolkit web components they
+// must be registered with the browser (i.e. webview) using the
+// syntax below.
+provideVSCodeDesignSystem().register(
+  vsCodeButton(),
+  vsCodeTextField(),
+  vsCodePanelView(),
+  vsCodeLink(),
+  vsCodeTag(),
+  vsCodeCheckbox(),
+  vsCodeDivider(),
+  vsCodeRadio(),
+  vsCodeRadioGroup(),
+  vsCodeProgressRing(),
+);
+
+bootstrapApplication(AppComponent).catch((err) => console.error(err));