+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Support for Angular 19 migration
 - Fix for yarn outdated plugin
 - Fix for Capacitor migrate with modern yarn
+- Fix for dialogs of "prepare release" losing focus
 
 ### Version 1.98